project/requirements.txt
altgraph==0.17.5
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2.11
django-auditlog==3.4.1
django-tailwind==4.4.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pefile==2024.8.26
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.3.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
